Post method working in server test (path directory fix)
