--- Develop/db/db.json
@@ -1,6 +1,0 @@
-[
-    {
-        "title":"Test Title",
-        "text":"Test text"
-    }
-]
