+++ db/db.json
@@ -1,0 +1,12 @@
+[
+    {
+        "title": "Test Title",
+        "text": "Test text",
+        "id": 1
+    },
+    {
+        "title": "Hey",
+        "text": "testtt",
+        "id": "5d37"
+    }
+]
